refactor: refactoring all classes on services folder
