+++ semana22/aula64/src/services/IdGenerator.ts
@@ -1,0 +1,7 @@
+import {v4 as uuid} from 'uuid';
+
+export default class IdGenerator {
+    generate(): string {
+        return uuid();
+    }
+}
